+++ standalone/smoketest.py
@@ -33,7 +33,7 @@ VIEWS = [
     "priorities_view", "committee_view", "hypotheses_view", "replication_view",
     "exp11_view", "integrity_view", "congress_view", "congress_reg_view",
     "government_view", "director_view", "decisions_view", "positions_view",
-    "market_summary",
+    "market_summary", "agents_view", "ops_view",
 ]
 
 failures = []